Inventory Update
This required us to compare two sets of arrays, inventory 1 and Inventory 2 and update the overall inventory.

This meant that if there was a match between an item in arr 1 and arr2, we had to add the total number of items together. If there was no match, we simply added the item into the overall data.

The hard part in this assignment was to sort it alphabetically.